requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
pyttsx3>=2.90
numpy>=1.24.0
